refactored AppDelegate and SceneDelegating by creating a CoordinatorManager reducing shred code
