+++ SwiftUIThingsILike/Sources/TabBar/TabBarCoordinator.swift
@@ -7,7 +7,6 @@
 
 import SwiftUI
 
-
 class TabBarCoordinator<Factory: TabBarViewFactoryType>: TabBarCoordinatorType {
     private let tabBarFactory: Factory
